+++ nodes/utility/destroy_and_recreate/destroy_and_recreate.py
@@ -61,5 +61,5 @@ NODE_CLASS_MAPPINGS = {
 }
 
 NODE_DISPLAY_NAME_MAPPINGS = {
-    "GoddessLabsDestroyAndRecreate": "Cache Buster ❤️‍🔥💊 GoddessLabs",
+    "GoddessLabsDestroyAndRecreate": "Destroy & Recreate ❤️‍🔥💊 GoddessLabs",
 }